Move announceBuild to new file
